requirements.txt
# Core dependencies
python-dotenv==1.0.0

# Document Processing
PyPDF2==3.0.1
pypdf==3.17.4
python-docx==1.1.0
langchain==0.1.0
langchain-community==0.0.10
langchain-text-splitters==0.0.1

# Encryption
cryptography==41.0.7
pycryptodome==3.19.0

# Embedding & Vector DB
sentence-transformers==2.2.2
qdrant-client==1.7.0

# LLM Integration
ollama==0.1.6

# Web Interface
fastapi==0.108.0
uvicorn[standard]==0.25.0
streamlit==1.29.0
python-multipart==0.0.6

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
numpy==1.24.3
pandas==2.1.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Monitoring & Logging
loguru==0.7.2
